Move min heap into separate subpackage
